+++ src/todoist.rs
@@ -1,13 +1,6 @@
-use std::env;
-
-use reqwest::blocking::Client;
-use reqwest::header::AUTHORIZATION;
-use reqwest::header::CONTENT_TYPE;
-use reqwest::header::USER_AGENT;
-use serde::Deserialize;
 use serde_json::json;
-use spinners::{Spinner, Spinners};
-use uuid::Uuid;
+
+mod request;
 
 use crate::config::Config;
 use crate::items::Item;
@@ -20,34 +13,16 @@ use crate::{items, projects, sections};
 const QUICK_ADD_URL: &str = "/sync/v9/quick/add";
 const PROJECT_DATA_URL: &str = "/sync/v9/projects/get_data";
 const SYNC_URL: &str = "/sync/v9/sync";
-const REST_V2_TASKS_URL: &str = "/rest/v2/tasks/";
+pub const REST_V2_TASKS_URL: &str = "/rest/v2/tasks/";
 const SECTIONS_URL: &str = "/rest/v2/sections";
 const PROJECTS_URL: &str = "/rest/v2/projects";
-
-// CRATES.IO URLS
-const VERSIONS_URL: &str = "/v1/crates/tod/versions";
-
-const FAKE_UUID: &str = "42963283-2bab-4b1f-bad2-278ef2b6ba2c";
-
-const SPINNER: Spinners = Spinners::Dots4;
-const MESSAGE: &str = "Querying API";
-
-#[derive(Deserialize)]
-struct CargoResponse {
-    versions: Vec<Version>,
-}
-
-#[derive(Deserialize)]
-struct Version {
-    num: String,
-}
 
 /// Add a new item to the inbox with natural language support
 pub fn quick_add_item(config: &Config, content: &str) -> Result<Item, String> {
     let url = String::from(QUICK_ADD_URL);
     let body = json!({"text": content, "auto_reminder": true});
 
-    let json = post_todoist_sync(config, url, body)?;
+    let json = request::post_todoist_sync(config, url, body)?;
     items::json_to_item(json)
 }
 
@@ -61,7 +36,7 @@ pub fn add_item(
     let url = String::from(REST_V2_TASKS_URL);
     let body = json!({"content": content, "description": description, "auto_reminder": true, "priority": priority.to_integer()});
 
-    let json = post_todoist_rest(config, url, body)?;
+    let json = request::post_todoist_rest(config, url, body)?;
     items::json_to_item(json)
 }
 
@@ -69,18 +44,18 @@ pub fn add_item(
 pub fn items_for_project(config: &Config, project_id: &str) -> Result<Vec<Item>, String> {
     let url = String::from(PROJECT_DATA_URL);
     let body = json!({ "project_id": project_id });
-    let json = post_todoist_sync(config, url, body)?;
+    let json = request::post_todoist_sync(config, url, body)?;
     items::json_to_items(json)
 }
 
 pub fn sections_for_project(config: &Config, project_id: &str) -> Result<Vec<Section>, String> {
     let url = format!("{SECTIONS_URL}?project_id={project_id}");
-    let json = get_todoist_rest(config, url)?;
+    let json = request::get_todoist_rest(config, url)?;
     sections::json_to_sections(json)
 }
 
 pub fn projects(config: &Config) -> Result<Vec<Project>, String> {
-    let json = get_todoist_rest(config, PROJECTS_URL.to_string())?;
+    let json = request::get_todoist_rest(config, PROJECTS_URL.to_string())?;
     projects::json_to_projects(json)
 }
 
@@ -91,10 +66,10 @@ pub fn move_item_to_project(
     project_name: &str,
 ) -> Result<String, String> {
     let project_id = projects::project_id(config, project_name)?;
-    let body = json!({"commands": [{"type": "item_move", "uuid": new_uuid(), "args": {"id": item.id, "project_id": project_id}}]});
+    let body = json!({"commands": [{"type": "item_move", "uuid": request::new_uuid(), "args": {"id": item.id, "project_id": project_id}}]});
     let url = String::from(SYNC_URL);
 
-    post_todoist_sync(config, url, body)?;
+    request::post_todoist_sync(config, url, body)?;
     Ok(String::from("✓"))
 }
 
@@ -103,10 +78,10 @@ pub fn move_item_to_section(
     item: Item,
     section_id: &str,
 ) -> Result<String, String> {
-    let body = json!({"commands": [{"type": "item_move", "uuid": new_uuid(), "args": {"id": item.id, "section_id": section_id}}]});
+    let body = json!({"commands": [{"type": "item_move", "uuid": request::new_uuid(), "args": {"id": item.id, "section_id": section_id}}]});
     let url = String::from(SYNC_URL);
 
-    post_todoist_sync(config, url, body)?;
+    request::post_todoist_sync(config, url, body)?;
     Ok(String::from("✓"))
 }
 
@@ -119,7 +94,7 @@ pub fn update_item_priority(
     let body = json!({ "priority": priority });
     let url = format!("{}{}", REST_V2_TASKS_URL, item.id);
 
-    post_todoist_rest(&config, url, body)?;
+    request::post_todoist_rest(&config, url, body)?;
     // Does not pass back an item
     Ok(String::from("✓"))
 }
@@ -129,7 +104,7 @@ pub fn update_item_due(config: &Config, item: Item, due_string: String) -> Resul
     let body = json!({ "due_string": due_string });
     let url = format!("{}{}", REST_V2_TASKS_URL, item.id);
 
-    post_todoist_rest(config, url, body)?;
+    request::post_todoist_rest(config, url, body)?;
     // Does not pass back an item
     Ok(String::from("✓"))
 }
@@ -139,17 +114,17 @@ pub fn update_item_name(config: &Config, item: Item, new_name: String) -> Result
     let body = json!({ "content": new_name });
     let url = format!("{}{}", REST_V2_TASKS_URL, item.id);
 
-    post_todoist_rest(config, url, body)?;
+    request::post_todoist_rest(config, url, body)?;
     // Does not pass back an item
     Ok(String::from("✓"))
 }
 
 /// Complete the last item returned by "next item"
 pub fn complete_item(config: &Config) -> Result<String, String> {
-    let body = json!({"commands": [{"type": "item_close", "uuid": new_uuid(), "temp_id": new_uuid(), "args": {"id": config.next_id}}]});
+    let body = json!({"commands": [{"type": "item_close", "uuid": request::new_uuid(), "temp_id": request::new_uuid(), "args": {"id": config.next_id}}]});
     let url = String::from(SYNC_URL);
 
-    post_todoist_sync(config, url, body)?;
+    request::post_todoist_sync(config, url, body)?;
 
     if !cfg!(test) {
         config.clone().clear_next_id().save()?;
@@ -159,169 +134,11 @@ pub fn complete_item(config: &Config) -> Result<String, String> {
     Ok(String::from("✓"))
 }
 
-/// Post to Todoist via sync API
-/// We use sync when we want natural languague processing.
-fn post_todoist_sync(
-    config: &Config,
-    url: String,
-    body: serde_json::Value,
-) -> Result<String, String> {
-    #[cfg(not(test))]
-    let todoist_url: String = "https://api.todoist.com".to_string();
-    #[cfg(not(test))]
-    let _placeholder = &config.mock_url;
-
-    #[cfg(test)]
-    let todoist_url: String = config.mock_url.clone().expect("Mock URL not set");
-
-    let request_url = format!("{todoist_url}{url}");
-    let token = &config.token;
-
-    let spinner = maybe_start_spinner(config);
-    let response = Client::new()
-        .post(request_url)
-        .header(CONTENT_TYPE, "application/json")
-        .header(AUTHORIZATION, format!("Bearer {token}"))
-        .json(&body)
-        .send()
-        .or(Err("Did not get response from server"))?;
-
-    maybe_stop_spinner(spinner);
-
-    if response.status().is_success() {
-        Ok(response.text().or(Err("Could not read response text"))?)
-    } else {
-        Err(format!("Error: {:#?}", response.text()))
-    }
-}
-
-/// Post to Todoist via REST api
-/// We use this when we want more options and don't need natural language processing
-fn post_todoist_rest(
-    config: &Config,
-    url: String,
-    body: serde_json::Value,
-) -> Result<String, String> {
-    #[cfg(not(test))]
-    let todoist_url: String = "https://api.todoist.com".to_string();
-
-    #[cfg(test)]
-    let todoist_url: String = config.mock_url.clone().expect("Mock URL not set");
-
-    let token = &config.token;
-
-    let request_url = format!("{todoist_url}{url}");
-    let authorization: &str = &format!("Bearer {token}");
-    let spinner = maybe_start_spinner(config);
-
-    let response = Client::new()
-        .post(request_url)
-        .header(CONTENT_TYPE, "application/json")
-        .header(AUTHORIZATION, authorization)
-        .header("X-Request-Id", new_uuid())
-        .json(&body)
-        .send()
-        .or(Err("Did not get response from server"))?;
-
-    maybe_stop_spinner(spinner);
-
-    if response.status().is_success() {
-        Ok(response.text().or(Err("Could not read response text"))?)
-    } else {
-        Err(format!("Error: {:#?}", response.text()))
-    }
-}
-
-// Combine get and post into one function
-/// Get Todoist via REST api
-fn get_todoist_rest(config: &Config, url: String) -> Result<String, String> {
-    #[cfg(not(test))]
-    let todoist_url: String = "https://api.todoist.com".to_string();
-
-    #[cfg(test)]
-    let todoist_url: String = config.mock_url.clone().expect("Mock URL not set");
-
-    let token = config.token.clone();
-
-    let request_url = format!("{todoist_url}{url}");
-    let authorization: &str = &format!("Bearer {token}");
-    let spinner = maybe_start_spinner(config);
-    let response = Client::new()
-        .get(request_url)
-        .header(CONTENT_TYPE, "application/json")
-        .header(AUTHORIZATION, authorization)
-        .send()
-        .or(Err("Did not get response from server"))?;
-
-    maybe_stop_spinner(spinner);
-
-    if response.status().is_success() {
-        Ok(response.text().or(Err("Could not read response text"))?)
-    } else {
-        Err(format!("Error: {:#?}", response.text()))
-    }
-}
-
-/// Get latest version number from Cargo.io
-pub fn get_latest_version(config: Config) -> Result<String, String> {
-    #[cfg(not(test))]
-    let cargo_url: String = "https://crates.io/api".to_string();
-    let _token = config.token;
-
-    #[cfg(test)]
-    let cargo_url: String = config.mock_url.expect("Mock URL not set");
-
-    let request_url = format!("{cargo_url}{VERSIONS_URL}");
-
-    let response = Client::new()
-        .get(request_url)
-        .header(USER_AGENT, "Tod")
-        .send()
-        .or(Err("Did not get response from server"))?;
-
-    if response.status().is_success() {
-        let cr: CargoResponse =
-            serde_json::from_str(&response.text().or(Err("Could not read response text"))?)
-                .or(Err("Could not serialize to CargoResponse"))?;
-        Ok(cr.versions.first().unwrap().num.clone())
-    } else {
-        Err(format!("Error: {:#?}", response.text()))
-    }
-}
-
-/// Create a new UUID, required for Todoist API
-fn new_uuid() -> String {
-    if cfg!(test) {
-        String::from(FAKE_UUID)
-    } else {
-        Uuid::new_v4().to_string()
-    }
-}
-
-fn maybe_start_spinner(config: &Config) -> Option<Spinner> {
-    match env::var("DISABLE_SPINNER") {
-        Ok(_) => None,
-        _ => {
-            if let Some(true) = config.spinners {
-                let sp = Spinner::new(SPINNER, MESSAGE.into());
-                Some(sp)
-            } else {
-                None
-            }
-        }
-    }
-}
-fn maybe_stop_spinner(spinner: Option<Spinner>) {
-    if let Some(mut sp) = spinner {
-        sp.stop();
-        print!("\x1b[2K\r");
-    };
-}
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::items::{DateInfo, Item};
-    use crate::{test, time, VERSION};
+    use crate::{test, time};
     use pretty_assertions::assert_eq;
 
     #[test]
@@ -473,24 +290,5 @@ mod tests {
         let response = update_item_due(&config, item, "today".to_string());
         mock.assert();
         assert_eq!(response, Ok(String::from("✓")));
-    }
-
-    #[test]
-
-    fn latest_version_works() {
-        let mut server = mockito::Server::new();
-        let mock = server
-            .mock("GET", "/v1/crates/tod/versions")
-            .with_status(200)
-            .with_header("content-type", "application/json")
-            .with_body(test::responses::versions())
-            .create();
-
-        let config = test::fixtures::config().mock_url(server.url());
-
-        let response = get_latest_version(config);
-        mock.assert();
-
-        assert_eq!(response, Ok(String::from(VERSION)));
     }
 }