+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add description flag to `tod task create`
+- Add ability to skip tasks in `project prioritize` and `project empty`
 
 ## 2023-05-31 v0.4.3